+++ manager-parent/manager-v1/src/main/java/com/redhat/service/smartevents/manager/v1/models/ManagedResourceV1.java
@@ -7,11 +7,12 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.MappedSuperclass;
 
+import com.redhat.service.smartevents.infra.core.exceptions.HasErrorInformation;
 import com.redhat.service.smartevents.infra.core.models.ManagedResourceStatus;
 import com.redhat.service.smartevents.manager.core.models.ManagedResource;
 
 @MappedSuperclass
-public class ManagedResourceV1 extends ManagedResource {
+public class ManagedResourceV1 extends ManagedResource implements HasErrorInformation {
 
     @Column(name = "status")
     @Enumerated(EnumType.STRING)
@@ -26,6 +27,12 @@ public class ManagedResourceV1 extends ManagedResource {
 
     @Column(name = "deletion_requested_at", columnDefinition = "TIMESTAMP")
     private ZonedDateTime deletionRequestedAt;
+
+    @Column(name = "error_id")
+    private Integer errorId;
+
+    @Column(name = "error_uuid")
+    private String errorUUID;
 
     public ManagedResourceStatus getStatus() {
         return status;
@@ -57,6 +64,24 @@ public class ManagedResourceV1 extends ManagedResource {
 
     public void setDeletionRequestedAt(ZonedDateTime deletionRequestedAt) {
         this.deletionRequestedAt = deletionRequestedAt;
+    }
+
+    @Override
+    public Integer getErrorId() {
+        return errorId;
+    }
+
+    public void setErrorId(Integer errorId) {
+        this.errorId = errorId;
+    }
+
+    @Override
+    public String getErrorUUID() {
+        return errorUUID;
+    }
+
+    public void setErrorUUID(String errorUUID) {
+        this.errorUUID = errorUUID;
     }
 
     public boolean isActionable() {
