+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/persistence/models/Condition.java
@@ -1,6 +1,7 @@
 package com.redhat.service.smartevents.manager.v2.persistence.models;
 
 import java.time.ZonedDateTime;
+import java.util.Objects;
 import java.util.UUID;
 
 import javax.persistence.Column;
@@ -109,5 +110,26 @@ public class Condition {
 
     public void setLastTransitionTime(ZonedDateTime lastTransitionTime) {
         this.lastTransitionTime = lastTransitionTime;
+    }
+
+    /*
+     * See: https://vladmihalcea.com/how-to-implement-equals-and-hashcode-using-the-jpa-entity-identifier/
+     * In the context of JPA equality, our id is our unique business key as we generate it via UUID.
+     */
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Condition condition = (Condition) o;
+        return id.equals(condition.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
     }
 }
